requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
starlette>=0.27.0
uvicorn>=0.24.0
httpx>=0.25.0
